Added readDirStreamWithPtr
This version of `readDirStreamWith` takes a pre-allocated pointer as an
additional argument that is used to store the pointer to the dirent. It
is useful when you want to read a lot of entries and want to safe a few
allocations since you can reuse the pointer for each call to
`readDirStreamWithPtr`.